# openapi.yaml
openapi: 3.0.1
info:
  title: Costume Votes API
  version: '1.0'
  description: API for managing costume party voting
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /api/me:
    get:
      summary: Get ballot information for a guest
      operationId: getBallot
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            description: Token from guest card (e.g. K7F2-9QX4)
      responses:
        '200':
          description: Ballot data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BallotResponse'
        '404':
          description: Unknown token
  /api/vote:
    post:
      summary: Cast or change a vote (or undo one)
      operationId: castVote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteRequest'
      responses:
        '200':
          description: Vote result with remaining count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
        '400':
          description: Unknown category
        '403':
          description: Voting closed or own costume
...
components:
  schemas:
    BallotResponse:
      type: object
      properties:
        guest:
          $ref: '#/components/schemas/GuestSummary'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryInfo'
        voting:
          $ref: '#/components/schemas/VotingStatus'
        votes_cast:
          type: integer
        remaining:
          type: integer
    VoteRequest:
      type: object
      required: [category, costume_id]
      properties:
        category:
          type: string
          enum: [beautiful, scary, couple]
        costume_id:
          type: integer
        undo:
          type: string
          enum: ['1']
    VoteResponse:
      type: object
      properties:
        ok:
          type: boolean
        votes_cast:
          type: integer
        remaining:
          type: integer
        ballot:
          $ref: '#/components/schemas/BallotSummary'
...
    VotingStatus:
      type: object
      properties:
        open:
          type: boolean
        reason:
          type: string
          enum: [not-opened, deadline-passed, closed-by-host]
        closesAt:
          type: string
          format: date-time
...
    GuestSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        token:
          type: string
        onboarded:
          type: boolean
        costume_id:
          type: integer
...
    CostumeSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        kind:
          type: string
          enum: [single, couple, group]
        members:
          type: integer
...
    BallotSummary:
      type: object
      properties:
        costumes:
          type: array
          items:
            $ref: '#/components/schemas/CostumeSummary'
        votes:
          type: object
          additionalProperties:
            type: object
            properties:
              costume_id:
                type: integer
    CategoryInfo:
      type: object
      properties:
        key:
          type: string
          enum: [beautiful, scary, couple]
        label:
          type: string
...